Refactor tile updates out of state.js
Updating a board is not unique to the State object and shouldn't be
part of the state.js file. This fits more in line with patterns.js
since it already is able to match and update regions.
