+++ src/state.js
@@ -1,8 +1,7 @@
-import { Board } from "./board.js";
-import { matcher } from "./matcher.js";
-import { patterns, applyTileUpdate } from "./patterns.js";
+import { applyPatternTileUpdates } from "./patterns.js";
 
 /**
+ * @typedef {import("./board.js").Board} Board
  * @typedef {import("./board.js").Point} Point
  * @typedef {import("./tile.js").Direction} Direction
  * @typedef {import("./tile.js").PlayerTile} PlayerTile
@@ -64,21 +63,21 @@ export class State {
    */
   setTile(pt, tile) {
     this.board.setTile(pt, tile);
-    this.addUpdatedTile(pt);
-    this.addUpdatedTile([pt[0] - 1, pt[1]]);
-    this.addUpdatedTile([pt[0] - 1, pt[1] - 1]);
-    this.addUpdatedTile([pt[0] - 1, pt[1] + 1]);
-    this.addUpdatedTile([pt[0] + 1, pt[1]]);
-    this.addUpdatedTile([pt[0] + 1, pt[1] - 1]);
-    this.addUpdatedTile([pt[0] + 1, pt[1] + 1]);
-    this.addUpdatedTile([pt[0], pt[1] - 1]);
-    this.addUpdatedTile([pt[0], pt[1] + 1]);
+    this.#addUpdatedTile(pt);
+    this.#addUpdatedTile([pt[0] - 1, pt[1]]);
+    this.#addUpdatedTile([pt[0] - 1, pt[1] - 1]);
+    this.#addUpdatedTile([pt[0] - 1, pt[1] + 1]);
+    this.#addUpdatedTile([pt[0] + 1, pt[1]]);
+    this.#addUpdatedTile([pt[0] + 1, pt[1] - 1]);
+    this.#addUpdatedTile([pt[0] + 1, pt[1] + 1]);
+    this.#addUpdatedTile([pt[0], pt[1] - 1]);
+    this.#addUpdatedTile([pt[0], pt[1] + 1]);
   }
 
   /**
    * @param {Point} pt
    */
-  addUpdatedTile(pt) {
+  #addUpdatedTile(pt) {
     if (
       this.board.isInBounds(pt) &&
       this.updatedTiles.every(
@@ -99,12 +98,35 @@ export class State {
     }
   }
 
-  clearJustUpdated() {
+  #clearJustUpdated() {
     for (let y = 0; y < this.board.height; ++y) {
       for (let x = 0; x < this.board.width; ++x) {
         this.board.getTile([x, y]).justUpdated = false;
       }
     }
+  }
+
+  applyUpdates() {
+    const updatedPoints = applyPatternTileUpdates(
+      this.board,
+      this.updatedTiles
+    );
+
+    this.updatedTiles = [];
+    this.#clearJustUpdated();
+    for (const pt of updatedPoints) {
+      this.#addUpdatedTile(pt);
+      this.#addUpdatedTile([pt[0] - 1, pt[1]]);
+      this.#addUpdatedTile([pt[0] - 1, pt[1] - 1]);
+      this.#addUpdatedTile([pt[0] - 1, pt[1] + 1]);
+      this.#addUpdatedTile([pt[0] + 1, pt[1]]);
+      this.#addUpdatedTile([pt[0] + 1, pt[1] - 1]);
+      this.#addUpdatedTile([pt[0] + 1, pt[1] + 1]);
+      this.#addUpdatedTile([pt[0], pt[1] - 1]);
+      this.#addUpdatedTile([pt[0], pt[1] + 1]);
+    }
+
+    return updatedPoints;
   }
 }
 
@@ -278,109 +300,6 @@ export function movePlayer(state, direction) {
       }
     }
   }
-
-  return updatedPoints;
-}
-
-/**
- * Sorts the points from bottom right to top left
- *
- * @param {Point[]} points
- */
-function reverseSortPoints(points) {
-  return [...points].sort((p1, p2) => {
-    if (p1[1] > p2[1]) {
-      return -1;
-    } else if (p1[1] < p2[1]) {
-      return 1;
-    } else if (p1[0] > p2[0]) {
-      return -1;
-    } else if (p1[0] < p2[0]) {
-      return 1;
-    }
-
-    return 0;
-  });
-}
-
-/**
- * Gets the 3x3 region centered at a given point
- *
- * @param {Board} board
- * @param {Point} pt
- */
-function getPointCenteredRegion(board, pt) {
-  return [
-    board.getTile([pt[0] - 1, pt[1] - 1]),
-    board.getTile([pt[0], pt[1] - 1]),
-    board.getTile([pt[0] + 1, pt[1] - 1]),
-    board.getTile([pt[0] - 1, pt[1]]),
-    board.getTile([pt[0], pt[1]]),
-    board.getTile([pt[0] + 1, pt[1]]),
-    board.getTile([pt[0] - 1, pt[1] + 1]),
-    board.getTile([pt[0], pt[1] + 1]),
-    board.getTile([pt[0] + 1, pt[1] + 1]),
-  ];
-}
-
-/**
- *
- * @param {State} state
- * @param {Point} pt
- * @param {TileUpdate[]} updates
- */
-function applyRegionUpdates(state, pt, updates) {
-  /** @type {Point[]} */
-  const updatedPoints = [];
-
-  for (let y = -1; y < 2; ++y) {
-    for (let x = -1; x < 2; ++x) {
-      const update = updates[(x + 1) + 3 * (y + 1)];
-
-      /** @type {Point} */
-      const currentPoint = [pt[0] + x, pt[1] + y];
-
-      if (update) {
-        updatedPoints.push(currentPoint);
-
-        state.setTile(
-          currentPoint,
-          applyTileUpdate(
-            state.board.getTile(currentPoint),
-            update
-          )
-        );
-      }
-    }
-  }
-
-  return updatedPoints;
-}
-
-/**
- * Applies updates to each tile that needs an update
- *
- * @param {State} state
- * @returns {Point[]} the points that were updated
- */
-export function applyPatternTileUpdates(state) {
-  const sortedUpdatedTiles = reverseSortPoints(state.updatedTiles);
-
-  state.updatedTiles = [];
-
-  /** @type {Point[]} */
-  const updatedPoints = [];
-  for (const point of sortedUpdatedTiles) {
-    const region = getPointCenteredRegion(state.board, point);
-    for (const [pattern, updates] of patterns) {
-      if (matcher(region, pattern)) {
-        updatedPoints.push(...applyRegionUpdates(state, point, updates));
-        break;
-      }
-    }
-  }
-
-  state.clearJustUpdated();
 
   return updatedPoints;
 }
